requirements.txt
openai
python-dotenv
gspread
oauth2client
pandas
pillow
ffmpeg-python 